Move static font assets out of runner
